menambahkan fitur check delete

+++ app/Http/Controllers/DataPelangganController.php
@@ -8,6 +8,8 @@ use Maatwebsite\Excel\Facades\Excel;
 use App\Http\Controllers\Controller;
 use App\Imports\DataPelangganImport;
 use App\Models\DataPelangganModel;
+use Illuminate\Support\Facades\Session;
+
 
 class DataPelangganController extends Controller
 {
@@ -32,29 +34,28 @@ class DataPelangganController extends Controller
     }
     public function import_excel(Request $request)
     {
-        // validasi
         $this->validate($request, [
             'file' => 'required|mimes:csv,xls,xlsx'
         ]);
-
-        // menangkap file excel
         $file = $request->file('file');
-
-        // membuat nama file unik
         $nama_file = rand() . $file->getClientOriginalName();
-
-        // upload ke folder file_siswa di dalam folder public
         $file->move('file_pelanggan', $nama_file);
-
-        // import data
         Excel::import(new DataPelangganImport, public_path('/file_pelanggan/' . $nama_file));
 
-        // alihkan halaman kembali
         return redirect('/inputpelanggan');
     }
-    public function hapusPelanggan($id){
-        $pelanggan = DataPelangganModel::find($id);
-        $pelanggan->delete();
-        return redirect('/inputpelanggan');
+    public function hapusPelanggan(){
+        $hapus_items = request('check');
+        if($hapus_items){
+            foreach($hapus_items as $hapus){
+                $pelanggan = DataPelangganModel::find($hapus);
+                $pelanggan->delete();
+            }
+            Session::flash('success_hapus', 'Data berhasil dihapus');
+            return redirect('/inputpelanggan');
+        }else{
+            Session::flash('error_hapus', 'Data gagal dihapus');
+            return redirect('/inputpelanggan');
+        }
     }
 }
